Update Azure SDK dependency

+++ src/IronPigeon.Dart/PostalService.cs
@@ -47,7 +47,7 @@ namespace IronPigeon.Dart
         /// <param name="bytesCopiedProgress">Progress feedback in terms of bytes uploaded.</param>
         /// <param name="cancellationToken">The cancellation token.</param>
         /// <returns>The asynchronous result.</returns>
-        public virtual Task<IReadOnlyCollection<Channel.NotificationPostedReceipt>> PostAsync(Message message, IProgress<int>? bytesCopiedProgress = null, CancellationToken cancellationToken = default(CancellationToken))
+        public virtual Task<IReadOnlyCollection<Channel.NotificationPostedReceipt>> PostAsync(Message message, IProgress<long>? bytesCopiedProgress = null, CancellationToken cancellationToken = default(CancellationToken))
         {
             Requires.NotNull(message, nameof(message));
